Fix OrdPSQ tree balancing logic. fixes #14
The OrdPSQ tree balancing logic would never run on a node with
a leaf directly below it, even when the other child is deeply
nested. This appears to be the result of an attempt at optimizing
the code in the original Hinze paper, but missing this case.

This patch reverts the balancing logic to what is described in the
paper.

This fixes a test case that apparently no one noticed was failing.